NO-TICKET: execution-engine: move tracking_copy into a dir
